Work on making PAPPL 1.2 compile against libcups3.

+++ pappl/subscription.h
@@ -116,7 +116,7 @@ typedef void (*pappl_event_cb_t)(pappl_system_t *system, pappl_printer_t *printe
 
 extern void		papplSubscriptionCancel(pappl_subscription_t *sub) _PAPPL_PUBLIC;
 
-extern pappl_subscription_t *papplSubscriptionCreate(pappl_system_t *system, pappl_printer_t *printer, pappl_job_t *job, int sub_id, pappl_event_t events, const char *username, const char *natural_language, const void *data, int datalen, int interval, int lease) _PAPPL_PUBLIC;
+extern pappl_subscription_t *papplSubscriptionCreate(pappl_system_t *system, pappl_printer_t *printer, pappl_job_t *job, int sub_id, pappl_event_t events, const char *username, const char *natural_language, const void *data, size_t datalen, int interval, int lease) _PAPPL_PUBLIC;
 
 extern pappl_event_t	papplSubscriptionGetEvents(pappl_subscription_t *sub) _PAPPL_PUBLIC;
 extern int		papplSubscriptionGetID(pappl_subscription_t *sub) _PAPPL_PUBLIC;
